move receipt related helpers to another file
